Fix: 카드 조회 에러 수정

+++ backend/src/main/java/com/ssafy/eureka/domain/statistics/dto/CardOwnershipDto.java
@@ -17,7 +17,7 @@ public class CardOwnershipDto {
     private String cardName;
     private String imagePath;
     private int imageAttr;
-    private int ownershipCount;
+    private Long ownershipCount;
 
     List<BenefitCompareInfo> benefitComapareList;
 
@@ -26,7 +26,7 @@ public class CardOwnershipDto {
         this.cardName = cardName;
         this.imagePath = imagePath;
         this.imageAttr = imageAttr;
-        this.ownershipCount = ownershipCount;
+        this.ownershipCount = (long) ownershipCount;
     }
 
     public CardOwnershipDto(int cardId, String cardName, String imagePath, int imageAttr, Long ownershipCount) {
@@ -34,7 +34,7 @@ public class CardOwnershipDto {
         this.cardName = cardName;
         this.imagePath = imagePath;
         this.imageAttr = imageAttr;
-        this.ownershipCount = ownershipCount.intValue();
+        this.ownershipCount = ownershipCount;
     }
 
     @NoArgsConstructor
